Tests decomposition
